+++ src/components/Dialogs/FileBrowser/FileBrowserDialogComponent.tsx
@@ -10,7 +10,7 @@ import {DraggableDialogComponent, TaskProgressDialogComponent} from "components/
 import {FileInfoComponent} from "components/FileInfo/FileInfoComponent";
 import {AppToaster, ErrorToast, type SimpleTableComponentProps} from "components/Shared";
 import {AnimationMode, BrowserMode, ColormapSet, DialogId, FileFilteringType, FileInfoType, HelpType, ImageType, PreferenceKeys} from "enums";
-import {NONLINEAR_SPECTRAL_AXIS_MESSAGES} from "models";
+import {NONLINEAR_SPECTRAL_AXIS_MESSAGE} from "models";
 import {AppStore, CatalogProfileStore, FileBrowserStore, type ISelectedFile, PreferenceStore} from "stores";
 import {type FrameStore} from "stores/Frame";
 
@@ -254,7 +254,7 @@ export class FileBrowserDialogComponent extends React.Component {
             throw new Error("No active frame");
         }
         if (activeFrame.isSpectralAxisNonlinear) {
-            throw new Error(NONLINEAR_SPECTRAL_AXIS_MESSAGES.saveImage);
+            throw new Error(NONLINEAR_SPECTRAL_AXIS_MESSAGE);
         }
 
         const saveFilename = fileBrowserStore.saveFilename;
@@ -625,7 +625,7 @@ export class FileBrowserDialogComponent extends React.Component {
                                     <small>To save color-blending images, please save as a workspace via the File menu.</small>
                                 </span>
                             ) : appStore.activeFrame?.isSpectralAxisNonlinear ? (
-                                <span>{NONLINEAR_SPECTRAL_AXIS_MESSAGES.saveImage}</span>
+                                <span>{NONLINEAR_SPECTRAL_AXIS_MESSAGE}</span>
                             ) : (
                                 "Save this file"
                             )